refactor create s3 function
