+++ bentoml/yatai/deployment/aws_lambda/operator.py
@@ -29,12 +29,12 @@ from bentoml.exceptions import (
 )
 from bentoml.saved_bundle import loader
 from bentoml.utils import status_pb_to_error_code_and_message
-from bentoml.utils.s3 import create_s3_bucket_if_not_exists
 from bentoml.utils.tempdir import TempDirectory
 from bentoml.yatai.deployment.aws_utils import (
     validate_sam_template,
     FAILED_CLOUDFORMATION_STACK_STATUS,
     cleanup_s3_bucket_if_exist,
+    create_s3_bucket_if_not_exists,
 )
 from bentoml.yatai.deployment.aws_lambda.utils import (
     init_sam_project,
